requirements.txt
asgiref==3.3.1
cffi==1.14.5
cryptography==3.4.6
Django==3.1.7
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework==3.12.2
pycparser==2.20
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
